requirements.txt
pandas
scikit-learn
streamlit
numpy
requests
python-dotenv
